moved filter to be under query.
